new file:   .metadata/.lock 	new file:   .metadata/.log 	new file:   .metadata/.mylyn/repositories.xml.zip 	new file:   .metadata/.plugins/com.genuitec.eclipse.devstyle/recent.json 	new file:   .metadata/.plugins/com.genuitec.eclipse.monitor/myeclipse-usage.properties 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/color-definitions-fix.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/colors-org.eclipse.buildship.ui-1231266899.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/colors-org.eclipse.egit.ui-1849137576.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/colors-org.eclipse.epp.mpc.ui.css--1027399938.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/colors-org.eclipse.jdt.ui-326703266.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/colors-org.eclipse.mylyn.tasks.ui--477129374.css 	new file:   .metadata/.plugins/com.genuitec.eclipse.ui.common.platform/splash-fix.css 	new file:   .metadata/.plugins/org.eclipse.core.resources/.root/.indexes/history.version 	new file:   .metadata/.plugins/org.eclipse.core.resources/.root/.indexes/properties.index 	new file:   .metadata/.plugins/org.eclipse.core.resources/.root/.indexes/properties.version 	new file:   .metadata/.plugins/org.eclipse.core.resources/.root/1.tree 	new file:   .metadata/.plugins/org.eclipse.core.resources/.safetable/org.eclipse.core.resources 	new file:   .metadata/.plugins/org.eclipse.e4.workbench/workbench.xmi 	new file:   .metadata/.plugins/org.eclipse.jdt.core/assumedExternalFilesCache 	new file:   .metadata/.plugins/org.eclipse.jdt.core/externalFilesCache 	new file:   .metadata/.plugins/org.eclipse.jdt.core/nonChainingJarsCache 	new file:   .metadata/.plugins/org.eclipse.jdt.core/variablesAndContainers.dat 	new file:   .metadata/.plugins/org.eclipse.jdt.ui/OpenTypeHistory.xml 	new file:   .metadata/.plugins/org.eclipse.jdt.ui/QualifiedTypeNameHistory.xml 	new file:   .metadata/.plugins/org.eclipse.m2e.logback.configuration/0.log 	new file:   .metadata/.plugins/org.eclipse.m2e.logback.configuration/logback.1.16.0.20200318-1040.xml 	new file:   .metadata/.plugins/org.eclipse.oomph.setup/workspace.setup 	new file:   .metadata/.plugins/org.eclipse.tips.ide/dialog_settings.xml 	new file:   .metadata/.plugins/org.eclipse.tm.terminal.view.ui/.executables/data.properties 	new file:   .metadata/.plugins/org.eclipse.ui.ide/dialog_settings.xml 	new file:   .metadata/.plugins/org.eclipse.ui.intro/introstate 	new file:   .metadata/.plugins/org.eclipse.ui.workbench/dialog_settings.xml 	new file:   .metadata/.plugins/org.eclipse.ui.workbench/workingsets.xml 	new file:   .metadata/defender.check 	new file:   .metadata/version.ini 	modified:   pom.xml 	new file:   src/main/java/vn/daoanhthanh/SE2_Hospital/Repository/patient/PatientDetailsRepository.java 	renamed:    src/main/java/vn/daoanhthanh/SE2_Hospital/Repository/PatientRepository.java -> src/main/java/vn/daoanhthanh/SE2_Hospital/Repository/patient/PatientRepository.java 	modified:   src/main/java/vn/daoanhthanh/SE2_Hospital/Se2HospitalApplication.java 	renamed:    src/main/java/vn/daoanhthanh/SE2_Hospital/controller/PatientController.java -> src/main/java/vn/daoanhthanh/SE2_Hospital/controller/patient/PatientController.java 	new file:   src/main/java/vn/daoanhthanh/SE2_Hospital/controller/patient/PatientDetailsController.java 	new file:   src/main/java/vn/daoanhthanh/SE2_Hospital/error/Patient/PatientDetailsNotFoundAdvice.java 	new file:   src/main/java/vn/daoanhthanh/SE2_Hospital/error/Patient/PatientDetailsNotFoundException.java 	modified:   src/main/java/vn/daoanhthanh/SE2_Hospital/error/Patient/PatientNotFoundAdvice.java 	modified:   src/main/java/vn/daoanhthanh/SE2_Hospital/model/patient/Patient.java 	modified:   src/main/java/vn/daoanhthanh/SE2_Hospital/model/patient/PatientDetails.java

+++ src/main/java/vn/daoanhthanh/SE2_Hospital/model/patient/PatientDetails.java
@@ -16,10 +16,10 @@ public class PatientDetails {
     private String bloodtype;
 
     @Column(name = "height")
-    private byte heigth;
+    private int heigth;
 
     @Column(name = "weight")
-    private byte weigth;
+    private int weigth;
 
     @OneToOne
     @MapsId
@@ -29,7 +29,7 @@ public class PatientDetails {
     public PatientDetails() {
     }
 
-    public PatientDetails(long id, String bloodtype, byte heigth, byte weigth, Patient patient) {
+    public PatientDetails(long id, String bloodtype, int heigth, int weigth, Patient patient) {
         this.id = id;
         this.bloodtype = bloodtype;
         this.heigth = heigth;
@@ -37,7 +37,7 @@ public class PatientDetails {
         this.patient = patient;
     }
 
-    public PatientDetails(String bloodtype, byte heigth, byte weigth, Patient patient) {
+    public PatientDetails(String bloodtype, int heigth, int weigth, Patient patient) {
         this.bloodtype = bloodtype;
         this.heigth = heigth;
         this.weigth = weigth;
@@ -60,19 +60,19 @@ public class PatientDetails {
         this.bloodtype = bloodtype;
     }
 
-    public byte getHeigth() {
+    public int getHeigth() {
         return heigth;
     }
 
-    public void setHeigth(byte heigth) {
+    public void setHeigth(int heigth) {
         this.heigth = heigth;
     }
 
-    public byte getWeigth() {
+    public int getWeigth() {
         return weigth;
     }
 
-    public void setWeigth(byte weigth) {
+    public void setWeigth(int weigth) {
         this.weigth = weigth;
     }
